+++ src/screens/AboutUsScreen.js
@@ -1,19 +1,23 @@
 import React from 'react';
 import { StatusBar } from 'expo-status-bar';
-import { Image, StyleSheet, Text, View } from 'react-native';
+import { Image, ScrollView, StyleSheet, Text, View } from 'react-native';
 
 import colors from '../../assets/colors/colors';
 
 const AboutUsScreen = () => {
     return(
-        <View style = {styles.container}>
-            <Text style = { styles.headerMain }>About Us</Text>
-            <StatusBar style = "auto" />
-            <Image source = {require('../../assets/adaptive-icon-cropped.png')} style = { styles.logo }/>
-            <Text style = { styles.headerSub }>VolunTrack</Text>
-            <Text style = { styles.text }>We are a non-profitable organization with a vision to allow high-school students to reach their potential when it comes to gaining experience and acquiring the skills and knowledge they need</Text>
-            <Image source = {require('../../assets/images/naassom-azevedo-Q_Sei-TqSlc-unsplash.png')} style = { styles.image }/>
-        </View>
+        // Made 'About Us' Screen Scrollable
+        <ScrollView> 
+            <View style = {styles.container}>
+                <Text style = { styles.headerMain }>About Us</Text>
+                <StatusBar style = "auto" />
+                <Image source = {require('../../assets/adaptive-icon-cropped.png')} style = { styles.logo }/>
+                <Text style = { styles.headerSub }>VolunTrack</Text>
+                <Text style = { styles.text }>We are a non-profitable organization with a vision to allow high-school students to reach their potential when it comes to gaining experience and acquiring the skills and knowledge they need</Text>
+                <Image source = {require('../../assets/images/naassom-azevedo-Q_Sei-TqSlc-unsplash.png')} style = { styles.image }/>
+            </View>
+        </ScrollView>
+        // Made 'About Us' Screen Scrollable
     );
 }
 
@@ -54,6 +58,7 @@ const styles = StyleSheet.create ({
     image: {
         width: 303,
         height: 240,
+        marginBottom: 40
     },
